refactor : remove service interface
