+++ src/main/kotlin/com/group4/ticketingservice/entity/User.kt
@@ -1,8 +1,9 @@
 package com.group4.ticketingservice.entity
 
 
+import com.group4.ticketingservice.dto.SignUpRequest
 import com.group4.ticketingservice.utils.Authority
-import com.group4.ticketingservice.controller.SignUpRequest
+
 import jakarta.persistence.*
 import org.springframework.security.core.GrantedAuthority
 import org.springframework.security.core.authority.SimpleGrantedAuthority
